Fix that clicks on book toolbar trigger page navigation when sticky
This was reported via a customer who emailed DHH. It was working fine in Safari and Firefox but somehow in Chrome, if you scrolled so that pages in the TOC were underneath the book toolbar and clicked a button (add pages, list/grid, move/delete) the click would pass through to the page underneath and trigger navigation to that page. Visually the toolbar correctly appeared as if it was stacked on top of the pages but it didn't behave that way. Bumping the z-index causes clicks to behave as expected.